Bytter ut ConceptForm helt med hook versjonen
