+++ src/containers/ConceptPage/ConceptForm/ConceptForm.jsx
@@ -6,7 +6,7 @@
  *
  */
 
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
@@ -29,63 +29,50 @@ import AccordionSection from './AccordionSection';
 import FormFooter from './FormFooter';
 import { ConceptShape, LicensesArrayOf, SubjectShape } from '../../../shapes';
 
-class ConceptForm extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      savedToServer: false,
-      translateOnContinue: false,
-    };
-    this.formik = React.createRef();
-  }
+const ConceptForm = ({
+  concept,
+  setConcept,
+  createMessage,
+  fetchConceptTags,
+  fetchStateStatuses,
+  inModal,
+  isNewlyCreated,
+  licenses,
+  onClose,
+  subjects,
+  translateConcept,
+  updateConceptAndStatus,
+  onUpdate,
+  applicationError,
+  t,
+}) => {
+  const [savedToServer, setSavedToServer] = useState(false);
+  const [translateOnContinue, setTranslateOnContinue] = useState(false);
 
-  componentDidUpdate({ concept: prevConcept }) {
-    const { concept } = this.props;
-    if (concept.language !== prevConcept.language || concept.id !== prevConcept.id) {
-      this.setState({ savedToServer: false });
-      if (this.formik.current) {
-        this.formik.current.resetForm();
-      }
-    }
-    this.formik = React.createRef();
-  }
+  useEffect(() => {
+    setSavedToServer(false);
+  }, [concept]);
+  const initialValues = transformApiConceptToFormValues(concept, subjects);
 
-  setTranslateOnContinue = translateOnContinue => {
-    this.setState({ translateOnContinue: translateOnContinue });
-  };
-
-  handleSubmit = async formik => {
+  const handleSubmit = async formik => {
     formik.setSubmitting(true);
-    const {
-      onUpdate,
-      concept,
-      applicationError,
-      updateConceptAndStatus,
-      setConcept,
-      licenses,
-    } = this.props;
-    const { revision } = concept;
-    const values = formik.values;
-    const initialValues = formik.initialValues;
-    const initialStatus = concept.status?.current;
-    const newStatus = formik.values.status?.current;
+    const { revision, status } = concept;
+    const { values, initialValues, errors } = formik;
+    const initialStatus = status?.current;
+    const newStatus = values.status?.current;
     const statusChange = initialStatus !== newStatus;
 
-    if (
-      formik.errors &&
-      Object.keys(formik.errors).length > 0 &&
-      formik.errors.constructor === Object
-    ) {
-      setConcept({ status: concept.status, ...getApiConcept(values, initialValues, licenses) });
+    if (errors && Object.keys(errors).length > 0 && errors.constructor === Object) {
+      setConcept({ status: status, ...getApiConcept(values, initialValues, licenses) });
       // if formik has errors, we stop submitting and show the error message(s)
-      const e = Object.keys(formik.errors).map(key => `${key}: ${formik.errors[key]}`);
-      this.props.createMessage({
+      const e = Object.keys(errors).map(key => `${key}: ${errors[key]}`);
+      createMessage({
         message: e.join(' '),
         severity: 'danger',
         timeToLive: 0,
       });
       formik.setSubmitting(false);
-      this.setState({ savedToServer: false });
+      setSavedToServer(false);
       return;
     }
 
@@ -112,109 +99,90 @@ class ConceptForm extends Component {
       }
       formik.resetForm();
       formik.setSubmitting(false);
-      this.setState({ savedToServer: true });
+      setSavedToServer(true);
     } catch (err) {
       applicationError(err);
       formik.setSubmitting(false);
-      this.setState({ savedToServer: false });
+      setSavedToServer(false);
     }
   };
 
-  render() {
-    const {
-      concept,
-      createMessage,
-      fetchConceptTags,
-      fetchStateStatuses,
-      inModal,
-      isNewlyCreated,
-      licenses,
-      onClose,
-      subjects,
-      t,
-      translateConcept,
-    } = this.props;
-    const { savedToServer, translateOnContinue } = this.state;
-
-    const initialValues = transformApiConceptToFormValues(concept, subjects);
-    return (
-      <Formik
-        initialValues={initialValues}
-        onSubmit={(values, actions) => {
-          this.handleSubmit({ ...actions, values });
-        }}
-        innerRef={this.formik}
-        enableReinitialize
-        validateOnMount
-        validate={values => validateFormik(values, conceptFormRules, t)}>
-        {formikProps => {
-          const { values, errors } = formikProps;
-          return (
-            <FormWrapper inModal={inModal} {...formClasses()}>
-              <HeaderWithLanguage
-                content={concept}
-                editUrl={lang => toEditConcept(values.id, lang)}
-                getEntity={() => getApiConcept(values, initialValues, licenses)}
-                translateArticle={translateConcept}
-                type="concept"
-                setTranslateOnContinue={this.setTranslateOnContinue}
-                values={values}
-              />
-              <AccordionWrapper>
-                <AccordionSection
-                  id="concept-content"
-                  title={t('form.contentSection')}
-                  className="u-4/6@desktop u-push-1/6@desktop"
-                  hasError={['title', 'conceptContent'].some(field => !!errors[field])}
-                  startOpen>
-                  <ConceptContent classes={formClasses} />
-                </AccordionSection>
-                <AccordionSection
-                  id="concept-copyright"
-                  title={t('form.copyrightSection')}
-                  className="u-6/6"
-                  hasError={['creators', 'license'].some(field => !!errors[field])}>
-                  <ConceptCopyright
-                    licenses={licenses}
-                    disableAgreements
-                    label={t('form.concept.source')}
-                  />
-                </AccordionSection>
-                <AccordionSection
-                  id="concept-metadataSection"
-                  title={t('form.metadataSection')}
-                  className="u-6/6"
-                  hasError={['tags', 'metaImageAlt', 'subjects'].some(field => !!errors[field])}>
-                  <ConceptMetaData fetchTags={fetchConceptTags} subjects={subjects} />
-                </AccordionSection>
-                <AccordionSection
-                  id="concept-articles"
-                  title={t('form.articleSection')}
-                  className="u-6/6"
-                  hasError={['articles'].some(field => !!errors[field])}>
-                  <ConceptArticles />
-                </AccordionSection>
-              </AccordionWrapper>
-              <FormFooter
-                entityStatus={concept.status}
-                inModal={inModal}
-                savedToServer={savedToServer}
-                isNewlyCreated={isNewlyCreated}
-                showSimpleFooter={!concept.id}
-                onClose={onClose}
-                onContinue={translateOnContinue ? translateConcept : () => {}}
-                handleSubmit={this.handleSubmit}
-                createMessage={createMessage}
-                getStateStatuses={fetchStateStatuses}
-                getApiConcept={() => getApiConcept(values, initialValues, licenses)}
-              />
-            </FormWrapper>
-          );
-        }}
-      </Formik>
-    );
-  }
-}
+  return (
+    <Formik
+      initialValues={initialValues}
+      onSubmit={(values, actions) => {
+        handleSubmit({ ...actions, values });
+      }}
+      enableReinitialize
+      validateOnMount
+      validate={values => validateFormik(values, conceptFormRules, t)}>
+      {formikProps => {
+        const { values, errors } = formikProps;
+        return (
+          <FormWrapper inModal={inModal} {...formClasses()}>
+            <HeaderWithLanguage
+              content={concept}
+              editUrl={lang => toEditConcept(values.id, lang)}
+              getEntity={() => getApiConcept(values, initialValues, licenses)}
+              translateArticle={translateConcept}
+              type="concept"
+              setTranslateOnContinue={setTranslateOnContinue}
+              values={values}
+            />
+            <AccordionWrapper>
+              <AccordionSection
+                id="concept-content"
+                title={t('form.contentSection')}
+                className="u-4/6@desktop u-push-1/6@desktop"
+                hasError={['title', 'conceptContent'].some(field => !!errors[field])}
+                startOpen>
+                <ConceptContent classes={formClasses} />
+              </AccordionSection>
+              <AccordionSection
+                id="concept-copyright"
+                title={t('form.copyrightSection')}
+                className="u-6/6"
+                hasError={['creators', 'license'].some(field => !!errors[field])}>
+                <ConceptCopyright
+                  licenses={licenses}
+                  disableAgreements
+                  label={t('form.concept.source')}
+                />
+              </AccordionSection>
+              <AccordionSection
+                id="concept-metadataSection"
+                title={t('form.metadataSection')}
+                className="u-6/6"
+                hasError={['tags', 'metaImageAlt', 'subjects'].some(field => !!errors[field])}>
+                <ConceptMetaData fetchTags={fetchConceptTags} subjects={subjects} />
+              </AccordionSection>
+              <AccordionSection
+                id="concept-articles"
+                title={t('form.articleSection')}
+                className="u-6/6"
+                hasError={['articles'].some(field => !!errors[field])}>
+                <ConceptArticles />
+              </AccordionSection>
+            </AccordionWrapper>
+            <FormFooter
+              entityStatus={concept.status}
+              inModal={inModal}
+              savedToServer={savedToServer}
+              isNewlyCreated={isNewlyCreated}
+              showSimpleFooter={!concept.id}
+              onClose={onClose}
+              onContinue={translateOnContinue ? translateConcept : () => {}}
+              handleSubmit={handleSubmit}
+              createMessage={createMessage}
+              getStateStatuses={fetchStateStatuses}
+              getApiConcept={() => getApiConcept(values, initialValues, licenses)}
+            />
+          </FormWrapper>
+        );
+      }}
+    </Formik>
+  );
+};
 
 ConceptForm.propTypes = {
   applicationError: PropTypes.func.isRequired,
