+++ src/containers/ConceptPage/CreateConcept.jsx
@@ -13,7 +13,7 @@ import { withRouter } from 'react-router-dom';
 import { HelmetWithTracker } from '@ndla/tracker';
 import { useFetchConceptData } from '../FormikForm/formikConceptHooks';
 import { toEditConcept } from '../../util/routeHelpers';
-import ConceptForm from './ConceptForm/ConceptHookForm';
+import ConceptForm from './ConceptForm';
 import { LicensesArrayOf } from '../../shapes';
 
 const CreateConcept = props => {
